added a few updates to the manual uploading/deleting of an entry

+++ app.py
@@ -4,7 +4,6 @@ from mongoengine.queryset.visitor import Q
 from resources import typeform, m_person
 
 app = Flask(__name__)
-app.config['UPLOAD_FOLDER'] = 'uploads/'
 app.config['MONGODB_SETTINGS'] = {
     'db': 'm_engine_db',
     'host': 'mongodb://tamuhack17:Tamuhackdb17@ds129600.mlab.com:29600/m_engine_db'
@@ -30,27 +29,28 @@ def home_page():
 # route that handles buttons on the home page
 @app.route("/modify", methods=['GET', 'POST'])
 def modify():
+    people = Person.objects()
     if request.method == 'POST':
         if request.form['action'] == 'add':
-            fname = request.form['fname_add']
-            lname = request.form['lname_add']
+            first_name = request.form['fname_add']
+            last_name = request.form['lname_add']
             email = request.form['email_add']
-            if fname and lname and email:
-                add = Person()
-                add.first_name = fname
-                add.last_name = lname
-                add.save()
-            count = Person.objects.count()
-            return render_template('add_delete.html', count=count)
+            save_dict = {"first_name": first_name, "last_name": last_name, "email": email}
+            if first_name and last_name and email:
+                num_uploads = parser.save_single(save_dict)
+            output_str = "Successfully Uploaded " + str(num_uploads)
+            return render_template('results.html', count=people.count(), entries=people, msg=output_str)
         elif request.form['action'] == 'delete':
-            query = request.form['fname_delete']
-            to_delete = Person.objects(first_name=query)
-            to_delete.delete()
-            count = Person.objects.count()
-            return render_template('add_delete.html', count=count)
+            query = request.form['email_delete']
+            to_delete = people(email=query)
+            num_delete = 0
+            if to_delete:
+                to_delete.delete()
+                num_delete = 1
+            output_str = "Successfully Deleted " + str(num_delete)
+            return render_template('results.html', count=people.count(), entries=people, msg=output_str)
     elif request.method == 'GET':
-        count = Person.objects.count()
-        return render_template('add_delete.html', count=count)
+        return render_template('add_delete.html', count=people.count())
 
 @app.route('/participants')
 def participants():
@@ -60,19 +60,20 @@ def participants():
     action = request.args.get('action', '')
     if action:
         if action == 'uploaded':
-            parser.save_data()
+            num_uploads = parser.save_group()
             count = Person.objects.count()
             entries = Person.objects()
-            return render_template('results.html', count=count, entries=entries, msg="Successful Upload")
+            output_str = "Successfully Uploaded " + str(num_uploads)
+            return render_template('results.html', count=count, entries=entries, msg=output_str)
         elif action == 'return':
-            return render_template('results.html', count=count, entries=entries, msg="You didn't upload")
+            return render_template('results.html', count=count, entries=entries, msg="No Upload")
     # otherwise, its a query or an empty query
     else:
         # second value is a default argument ''
         query = request.args.get('q', '')
         if query:
             entries = Person.objects(Q(first_name__icontains=query) | Q(last_name__icontains=query) | Q(email__icontains=query))
-        return render_template('results.html', count=count, entries=entries, query=query, msg="test message")
+        return render_template('results.html', count=count, entries=entries, query=query, msg="Displaying Search Results")
 
 @app.route("/upload", methods=['GET', 'POST'])
 def upload():
